// src/api/generated/model/hourPreferenceDataType.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API
 * API for the Ludora app
 * OpenAPI spec version: 0.0.1
 */

/**
 * The preference type, RECURRENT is a recurring preference, ONE_TIME is a one time preference
 */
export type HourPreferenceDataType = (typeof HourPreferenceDataType)[keyof typeof HourPreferenceDataType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const HourPreferenceDataType = {
  RECURRENT: 'RECURRENT',
  ONE_TIME: 'ONE_TIME',
} as const;